+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 3.1.1 (2023.5.xx)
+
+[Release Link](https://github.com/Nexusoft/NexusInterface/releases/tag/v3.1.1)
+
+#### Additions
+
+- Added 5.1 Documents to the Terminal's API menu
+
+#### Adjustments
+
+-
+
+#### Fixes
+
+- Fixed crash when clicking the Send All function on the Send page
+- Fixed the Terminal's API Doc display
+
 # 3.1.0 (2023.4.26)
 
 [Release Link](https://github.com/Nexusoft/NexusInterface/releases/tag/v3.1.0)